Command line options: battery calibration test
Not a whole lot to there. This is just a video that demonstrates the program works. For each and every setting to work, additional modifications must be made to the config file. I likely won't finish tweaking everything but for now this works well enough for people who hope to enjoy the product's core functionality.