Create apcusd-config
More to come soon. This script should work - but I have not been able to test it on bare metal yet. I can confirm that my configuration file - which you can review in this repo works quite well. 

If you need this to work properly some of these commands may need to be run manually - and some trial and error may be necessary. If this script works for you, great. I did spend a lot of time working on it. 

Anyone who has questions, feel free to reach out.